JIT tuning; set cache size on command line
The tuning knobs for triggering trace compilation for the JIT
had not been revisited for several years.  In that time, the
working set of some applications have significantly increased,
leading to frequent cache overlows & flushes.

This CL adds the ability to set the maximum size of the JIT's
cache on the command line, and we expect to use different settings
depending on device configuration (rule of thumb: 1K for each 1M
for system RAM, with 2M limit).

Additionally, the trace compilation trigger has been tightened to
limit the compilation of cold traces.

Change-Id: Ic338d7dc1a4e8ad194a52ef964b133a56a2e170e